Custom Bots/Real Estate Lead Gen Bot/requirements.txt
# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
requests-html==0.10.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Location Services
uszipcode==1.0.1

# Web framework
flask==3.0.0
flask-sqlalchemy==3.1.1

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
requests-cache==1.1.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

DATABASE_URL=sqlite:///data/properties.db
GOOGLE_MAPS_API_KEY=your_google_maps_api_key
ZILLOW_API_KEY=your_zillow_api_key 